#58 Improved flow and error handling in calculator

+++ MathLibrary/Data/CalculationResult.cs
@@ -7,17 +7,19 @@ namespace ServiceLibrary.Data
         [Key]
         public int Id { get; set; }
         [Required]
-        public int FirstNumber { get; set; }
+        public double FirstNumber { get; set; }
         [Required]
         public string Operator { get; set; }
         [Required]
-        public int SecondNumber { get; set; }
+        public double SecondNumber { get; set; }
         [Required]
-        public int EquationResult { get; set; }
+        public double EquationResult { get; set; }
         [Required]
         public DateTime DateOfCalculation { get; set; }
         [Required]
         public bool IsActive { get; set; }
+
+
 
     }
 }
